LayoutTests/platform/win/imported/w3c/web-platform-tests/trusted-types/trusted-types-reporting-expected.txt

CONSOLE MESSAGE: Refused to create a TrustedTypePolicy named 'three' because it violates the following Content Security Policy directive: "trusted-types one"
CONSOLE MESSAGE: [Report Only] Refused to create a TrustedTypePolicy named 'three' because it violates the following Content Security Policy directive: "trusted-types two"
CONSOLE MESSAGE: Refused to connect to http://web-platform.test:8800/common/blank.html because it does not appear in the connect-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to create a TrustedTypePolicy named 'two' because it violates the following Content Security Policy directive: "trusted-types one"
CONSOLE MESSAGE: Refused to connect to http://web-platform.test:8800/common/blank.html because it does not appear in the connect-src directive of the Content Security Policy.
CONSOLE MESSAGE: [Report Only] Refused to create a TrustedTypePolicy named 'one' because it violates the following Content Security Policy directive: "trusted-types two"
CONSOLE MESSAGE: Refused to connect to http://web-platform.test:8800/common/blank.html because it does not appear in the connect-src directive of the Content Security Policy.
CONSOLE MESSAGE: This requires a TrustedScriptURL value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: Refused to connect to http://web-platform.test:8800/common/blank.html because it does not appear in the connect-src directive of the Content Security Policy.
CONSOLE MESSAGE: This requires a TrustedScriptURL value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: Refused to connect to http://web-platform.test:8800/common/blank.html because it does not appear in the connect-src directive of the Content Security Policy.
CONSOLE MESSAGE: This requires a TrustedScript value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: Refused to connect to http://web-platform.test:8800/common/blank.html because it does not appear in the connect-src directive of the Content Security Policy.
CONSOLE MESSAGE: This requires a TrustedScriptURL value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: Refused to connect to http://web-platform.test:8800/common/blank.html because it does not appear in the connect-src directive of the Content Security Policy.
CONSOLE MESSAGE: This requires a TrustedScriptURL value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: Refused to connect to http://web-platform.test:8800/common/blank.html because it does not appear in the connect-src directive of the Content Security Policy.

PASS Trusted Type violation report: creating a forbidden policy.
FAIL Trusted Type violation report: creating a report-only-forbidden policy. assert_true: expected true got false
PASS Trusted Type violation report: creating a forbidden-but-not-reported policy.
PASS Trusted Type violation report: sample for SVGScriptElement href assignment
PASS Trusted Type violation report: sample for SVGScriptElement href assignment by setAttribute
PASS Trusted Type violation report: sample for eval
PASS Trusted Type violation report: sample for custom element assignment
PASS Trusted Type violation report: Worker constructor
